Initial commit.
Fellipe, Kate,

I'd appreciate it if one of you could review this code :)
This code is running on:
http://codelab-maas.appspot.com/

Initial commit.

This is supposed to be the initial state of the Codelab 'Memes as a Service'.

README is still empty, but I'll work on another CL.

Reviewed in https://codereview.appspot.com/13191043/